split Player and NetworkPlayer
